Fin de Seccion 10, Todo List terminado
